+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         <activity android:name=".room.close.RoomCloseActivity">
         </activity>
 
-        <activity android:name=".room.mypage.RoomMyPageActivity"/>
+        <activity android:name=".matching.home.MatchingHomeActivity"/>
 
         <activity android:name=".room.home.RoomHomeActivity">
         </activity>
