+++ app/src/main/java/com/mashup/frienitto/FrenttoMudule.kt
@@ -1,11 +1,11 @@
 package com.mashup.frienitto
 
 import com.mashup.frienitto.login.LoginViewModel
+import com.mashup.frienitto.matching.home.MatchingHomeViewModel
 import com.mashup.frienitto.register.RegisterFragmentViewModel
 import com.mashup.frienitto.register.RegisterViewModel
 import com.mashup.frienitto.repository.room.RoomRepository
 import com.mashup.frienitto.room.close.RoomCloseViewModel
-import com.mashup.frienitto.room.mypage.RoomMyPageViewModel
 import com.mashup.frienitto.room.creation.RoomCreationViewModel
 import com.mashup.frienitto.room.join.RoomJoinViewModel
 
@@ -23,7 +23,7 @@ val module: Module = module {
     viewModel { RegisterFragmentViewModel() }
     viewModel { RoomJoinViewModel(get(), get()) }
     viewModel { RoomHomeViewModel(get()) }
-    viewModel { RoomMyPageViewModel() }
+    viewModel { MatchingHomeViewModel(get()) }
     viewModel { RoomCloseViewModel() }
     single { RoomRepository() }
 }
